% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{extract_tabular_metadata}
\alias{extract_tabular_metadata}
\title{Extract Tabular Metadata}
\usage{
extract_tabular_metadata(df, id_field = NULL, spatial_unit = NULL)
}
\arguments{
\item{df}{A data frame}

\item{id_field}{Optional name of the ID field}

\item{spatial_unit}{Optional description of what each row represents}
}
\value{
A list of metadata fields
}
\description{
Extracts metadata from a data frame for the envelope.
}
