% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{extract_vector_metadata}
\alias{extract_vector_metadata}
\title{Extract Vector Metadata}
\usage{
extract_vector_metadata(sf_obj, id_field = NULL)
}
\arguments{
\item{sf_obj}{An sf object}

\item{id_field}{Optional name of the ID field}
}
\value{
A list of metadata fields
}
\description{
Extracts metadata from an sf object for the envelope.
}
