% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{get_param}
\alias{get_param}
\title{Get Parameter with Default}
\usage{
get_param(params, name, default = NULL)
}
\arguments{
\item{params}{The params list from parse_primitive_args()}

\item{name}{The parameter name}

\item{default}{Default value if not provided}
}
\value{
The parameter value
}
\description{
Get Parameter with Default
}
