% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R, R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{Null coalescing operator}
\usage{
a \%||\% b

a \%||\% b
}
\arguments{
\item{a}{First value}

\item{b}{Default value if a is NULL}
}
\value{
a if not NULL, otherwise b
}
\description{
Null coalescing operator

Null Coalescing Operator
}
\keyword{internal}
