% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{primitive_success}
\alias{primitive_success}
\title{Report Primitive Success}
\usage{
primitive_success(metadata, warnings = NULL)
}
\arguments{
\item{metadata}{List of metadata fields}

\item{warnings}{Warnings collector or list of warnings}
}
\description{
Outputs metadata as JSON to stdout and exits successfully.
Call this at the end of a successful primitive execution.
}
