% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_read_rast}
\alias{safe_read_rast}
\title{Safely Read Raster Data}
\usage{
safe_read_rast(path, warnings = NULL)
}
\arguments{
\item{path}{Path to file}

\item{warnings}{Warnings collector}
}
\value{
SpatRaster object
}
\description{
Reads raster data with error handling.
}
