% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_read_sf}
\alias{safe_read_sf}
\title{Safely Read Vector Data}
\usage{
safe_read_sf(path, warnings = NULL)
}
\arguments{
\item{path}{Path to file}

\item{warnings}{Warnings collector}
}
\value{
sf object
}
\description{
Reads spatial vector data with error handling.
}
