% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_write_rast}
\alias{safe_write_rast}
\title{Safely Write Raster Data}
\usage{
safe_write_rast(obj, path, warnings = NULL)
}
\arguments{
\item{obj}{SpatRaster object}

\item{path}{Output path}

\item{warnings}{Warnings collector}
}
\description{
Writes raster data with error handling.
}
