% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_write_sf}
\alias{safe_write_sf}
\title{Safely Write Vector Data}
\usage{
safe_write_sf(obj, path, warnings = NULL)
}
\arguments{
\item{obj}{sf object}

\item{path}{Output path}

\item{warnings}{Warnings collector}
}
\description{
Writes spatial vector data with error handling.
}
