% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_spatial_overlap}
\alias{validate_spatial_overlap}
\title{Validate Spatial Overlap}
\usage{
validate_spatial_overlap(obj1, obj2, warnings = NULL, min_overlap_pct = 0)
}
\arguments{
\item{obj1}{First spatial object}

\item{obj2}{Second spatial object}

\item{warnings}{Warnings collector}

\item{min_overlap_pct}{Minimum overlap percentage to consider valid}
}
\value{
TRUE if sufficient overlap, FALSE otherwise
}
\description{
Checks if two spatial objects overlap.
}
