% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{warnings_collector}
\alias{warnings_collector}
\title{Create a Warnings Collector}
\usage{
warnings_collector()
}
\value{
A warnings collector (list with functions)
}
\description{
Creates a mutable warnings list that can accumulate warnings
throughout primitive execution.
}
\examples{
w <- warnings_collector()
w$add("info", "CRS was transformed")
w$warn("3 features excluded")
w$get()
}
